add Lesson10

+++ Lesson10/src/main/java/by/teachmeskills/person/jacket/IJacket.java
@@ -1,10 +1,13 @@
 package by.teachmeskills.person.jacket;
 
+import java.math.BigDecimal;
+
 public interface IJacket {
 
     void putOn();
 
     void takeOff();
 
-    int getPrice();
+    BigDecimal getPrice();
+
 }
