+++ Lesson10/src/main/java/by/teachmeskills/person/shoes/IShoes.java
@@ -1,10 +1,12 @@
 package by.teachmeskills.person.shoes;
 
+import java.math.BigDecimal;
+
 public interface IShoes {
 
     void putOn();
 
     void takeOff();
 
-    int getPrice();
+    BigDecimal getPrice();
 }
